+++ src/app/routes/health.ts
@@ -1,9 +1,9 @@
 /**
  * ヘルスチェックルーター
  */
-import * as pecorino from '@pecorino/domain';
 import * as express from 'express';
 import { OK } from 'http-status';
+import * as mongoose from 'mongoose';
 
 const healthRouter = express.Router();
 
@@ -11,7 +11,7 @@ healthRouter.get(
     '',
     async (_, res, next) => {
         try {
-            await pecorino.mongoose.connection.db.admin().ping();
+            await mongoose.connection.db.admin().ping();
             res.status(OK).send('healthy!');
         } catch (error) {
             next(error);