+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - MongoDBインデックス調整
+- install @pecorino/domain@2.0.0
 
 ### Deprecated